Removed footer + added delete perm check

+++ interactions/delete.js
@@ -6,14 +6,19 @@ module.exports = {
 		.setDescription('Delete Messages')
     .addIntegerOption(option => option.setName("number").setDescription("The number of messages").setRequired(true)),
 	async execute(interaction) {
-    const Embed = new MessageEmbed().setTitle("Delete Messages").setTimestamp()
-    .setFooter(`${interaction.user.username}#${interaction.user.discriminator}` );
+    const Embed = new MessageEmbed().setTitle("Delete Messages").setTimestamp();
     const count = interaction.options.getInteger('number');
     if (count<=0 || count >99) Embed.setDescription("Invalid Number").setColor("#ff0000")
     else if (interaction.member.permissions.has(Permissions.FLAGS.MANAGE_MESSAGES))
        {
-       interaction.channel.bulkDelete(count);	
-        Embed.setDescription(`${count} messages deleted`).setColor("#0f3c6c");
+         if (interaction.guild.me.permissionsIn(interaction.channel).has(Permissions.FLAGS.MANAGE_MESSAGES))
+         {
+          interaction.channel.bulkDelete(count);	
+          Embed.setDescription(`${count} messages deleted`).setColor("#0f3c6c");
+         }
+         else {
+          Embed.setDescription(`I don't have permission to delete messages here!`).setColor("#ff0000")
+         }
       }
     else {Embed.setDescription("You do not have permission to use this command!").setColor("#ff0000");}
     interaction.reply({ embeds: [Embed], ephemeral: true})
